Created a class to represent the game and a module for each method that's used to run the app.
Find_longest_sequence contains a method that checks the initial array and searches for the longest sequence. It also uses recursion to return the check_new_sequence method from the module with the same name.

Check_new_sequence contains a method that is used to verify if a new sequence is created after removing the longest one. It also uses recursion to return the find_longest_sequence method if the length of the array allows for a new sequence to exist.

Output_message is a method used to print the results to the user based on the data gathered and validated by the first 2 methods.